backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3
pypdf==3.17.4
docx2txt==0.8
langchain==0.1.0
langchain-openai==0.0.5
langchain-google-genai==0.0.6
langchain-community==0.0.10
langchain-core==0.1.7
langchain-text-splitters==0.0.1
openai==1.3.7
google-generativeai==0.3.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.36.0
numpy==1.24.3
firebase-admin==6.2.0
google-cloud-firestore==2.13.1
httpx==0.25.2
aiofiles==23.2.1
pandas==2.1.4
scikit-learn==1.3.2
torch==2.1.1
torchvision==0.16.1
protobuf==4.25.1